faster_rcnn/requirements.txt
pillow
matplotlib
torch
torchvision
torchaudio
torchmetrics
scikit-learn
albumentations
pycocotools